chore(init): Reorganized init files + export constants
